# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sandbox administration and multi-tenant provisioning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/domains/{hostname}:
    get:
      operationId: resolvePublicDomain
      tags: [website]
      parameters: [{name: hostname, in: path, required: true, schema: {type: string, maxLength: 253}}]
      responses:
        "200":
          description: Active entitled tenant for a verified domain
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PublicSite"}
...
  /public/sites/{slug}/branding/{kind}:
    get:
      operationId: getPublicBrandingFile
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
        - {name: kind, in: path, required: true, schema: {type: string, enum: [logo, favicon]}}
      responses:
        "200":
          description: Curated raster logo or favicon explicitly selected for this delivered tenant. Other order attachments remain private.
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
        "404": {description: Branding file unavailable}
  /public/sites/{slug}:
    get:
      operationId: getPublicSite
      tags: [catalog]
      parameters: [{name: slug, in: path, required: true, schema: {type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"}}]
      responses:
        "200": {description: Allowlisted public tenant website configuration, content: {application/json: {schema: {$ref: "#/components/schemas/PublicSite"}}}}
  /public/sites/{slug}/capabilities/{feature}:
    get:
      operationId: getPublicCapability
      tags: [catalog]
      parameters: [{name: slug, in: path, required: true, schema: {type: string}}, {name: feature, in: path, required: true, schema: {type: string}}]
      responses:
        "200": {description: Enabled foundation entry point; no product execution, content: {application/json: {schema: {$ref: "#/components/schemas/CapabilityStatus"}}}}
components:
...
  schemas:
...
    AssetNetwork:
      type: object
      properties:
        assetId: { type: string }
        assetNetworkId: {type: string}
        symbol: { type: string }
        name: { type: string }
        networkId: { type: string }
        networkName: { type: string }
        testnet: { type: boolean }
        logoUrl: {type: ["string", "null"]}
        decimals: {type: integer}
      required: [assetId, symbol, name, networkId, networkName, testnet]
...
    WebsiteSettings:
      type: object
      properties:
        navigation:
          type: array
          maxItems: 12
          items:
            type: object
            properties:
              key: {type: string, enum: [exchange, payments, telegram, how, developers, about, faq]}
              label: {type: string, minLength: 1, maxLength: 60}
              visible: {type: boolean}
            required: [key, label, visible]
        secondaryColor: {type: string, pattern: "^#[0-9A-Fa-f]{6}$"}
        templateKey: {type: string, enum: [standard-exchange]}
        websiteName: {type: string, minLength: 2, maxLength: 120}
        sandboxLabel: {type: string, enum: [Sandbox, Sandbox Demo]}
        glowColor: {type: string, pattern: "^#[0-9A-Fa-f]{6}$"}
        surfaceStyle: {type: string, enum: [solid, glass]}
        borderRadius: {type: string, enum: [sharp, soft, rounded]}
        faq: {type: array, maxItems: 16, items: {$ref: "#/components/schemas/WebsiteFaq"}}
        faviconUrl: {type: ["string", "null"], maxLength: 2048}
        fontKey: {type: string, enum: [system, inter, manrope, dm-sans, space-grotesk]}
        heroTitle: {type: string, minLength: 2, maxLength: 180}
        heroSubtitle: {type: string, maxLength: 1000}
        supportEmail: {type: ["string", "null"], maxLength: 254}
        supportUrl: {type: ["string", "null"], maxLength: 2048}
        supportDetails: {type: string, maxLength: 2000}
        socialLinks: {type: array, maxItems: 12, items: {$ref: "#/components/schemas/SiteLink"}}
        footerText: {type: string, maxLength: 2000}
        privacyContent: {type: string, maxLength: 20000}
        termsContent: {type: string, maxLength: 20000}
      required: [secondaryColor, faviconUrl, fontKey, heroTitle, heroSubtitle, supportEmail, supportUrl, supportDetails, socialLinks, footerText, privacyContent, termsContent]
    WebsiteFaq:
      type: object
      properties:
        question: {type: string, minLength: 2, maxLength: 180}
        answer: {type: string, minLength: 2, maxLength: 2000}
      required: [question, answer]
    SiteLink:
      type: object
      properties: {label: {type: string, minLength: 1, maxLength: 80}, url: {type: string, maxLength: 2048}}
      required: [label, url]
...
    PublicSite:
      type: object
      properties:
        tenantSlug: {type: string}
        brandName: {type: string}
        logoUrl: {type: ["string", "null"]}
        primaryColor: {type: string}
        accentColor: {type: string}
        themeMode: {type: string, enum: [light, dark, system]}
        domain: {type: ["string", "null"]}
        sandboxOnly: {type: boolean}
        websiteSettings: {$ref: "#/components/schemas/WebsiteSettings"}
        features: {type: object, additionalProperties: {type: boolean}}
        assets: {type: array, items: {$ref: "#/components/schemas/AssetNetwork"}}
      required: [tenantSlug, brandName, logoUrl, primaryColor, accentColor, themeMode, domain, sandboxOnly, websiteSettings, features, assets]
    CapabilityStatus:
      type: object
      properties: {feature: {type: string}, status: {type: string}, message: {type: string}}
      required: [feature, status, message]